backend/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
coverage==7.4.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
executing==2.0.1
i==2.1.0
iniconfig==2.0.0
mysqlclient==2.2.3
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.0
python-decouple==3.8
pytz==2024.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
gunicorn==21.2.0
daphne==3.0.2
channels==3.0.4


